fix: cleaning code

+++ src/services/client/deleteClient.service.ts
@@ -1,11 +1,14 @@
 import { AppDataSource } from "../../data-source"
 import { Client } from "../../entities/client.entity"
-import { Contact } from "../../entities/contact.entity"
-import { clienteSchemaWithoutPassword, contactSchemaResponse, listClient } from "../../serializers/serializers"
+import AppError from "../../errors/AppError"
 
 export const deleteClientService = async (contactId:string) => {
     const clientRepo = AppDataSource.getRepository(Client)
     const client = await clientRepo.findOneBy({id:contactId})
+
+    if(!client){
+        throw new AppError("Client does not exists", 404)
+    }
 
     await clientRepo.remove(client!)
